rename: user 디렉토리 이동
