+++ server/src/main/java/challenge/server/bookmark/service/BookmarkService.java
@@ -7,9 +7,9 @@ import challenge.server.exception.ExceptionCode;
 import challenge.server.habit.entity.Habit;
 import challenge.server.habit.repository.HabitRepository;
 import challenge.server.habit.service.HabitService;
-import challenge.server.security.user.entity.User;
-import challenge.server.security.user.repository.UserRepository;
-import challenge.server.security.user.service.UserService;
+import challenge.server.user.entity.User;
+import challenge.server.user.repository.UserRepository;
+import challenge.server.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
