+++ server/src/main/java/challenge/server/config/SecurityConfig.java
@@ -10,8 +10,8 @@ import challenge.server.security.handler.UserAuthenticationFailureHandler;
 import challenge.server.security.handler.UserAuthenticationSuccessHandler;
 import challenge.server.security.jwt.JwtTokenizer;
 import challenge.server.security.utils.CustomAuthorityUtils;
-import challenge.server.security.user.service.LogoutListService;
-import challenge.server.security.user.service.UserService;
+import challenge.server.user.service.LogoutListService;
+import challenge.server.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
