+++ server/src/main/java/challenge/server/helper/event/UserRegistrationApplicationEvent.java
@@ -1,6 +1,6 @@
 package challenge.server.helper.event;
 
-import challenge.server.security.user.entity.User;
+import challenge.server.user.entity.User;
 import lombok.Getter;
 
 @Getter
