+++ server/src/main/java/challenge/server/report/service/ReportService.java
@@ -4,8 +4,8 @@ import challenge.server.exception.BusinessLogicException;
 import challenge.server.exception.ExceptionCode;
 import challenge.server.report.entity.Report;
 import challenge.server.report.repository.ReportRepository;
-import challenge.server.security.user.entity.User;
-import challenge.server.security.user.repository.UserRepository;
+import challenge.server.user.entity.User;
+import challenge.server.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
