+++ server/src/main/java/challenge/server/review/entity/Review.java
@@ -2,7 +2,7 @@ package challenge.server.review.entity;
 
 import challenge.server.audit.BaseTimeEntity;
 import challenge.server.habit.entity.Habit;
-import challenge.server.security.user.entity.User;
+import challenge.server.user.entity.User;
 import lombok.*;
 
 import javax.persistence.*;
