+++ server/src/main/java/challenge/server/security/controller/SecurityController.java
@@ -1,7 +1,7 @@
 package challenge.server.security.controller;
 
 import challenge.server.security.service.SecurityService;
-import challenge.server.security.user.mapper.UserMapperImpl;
+import challenge.server.user.mapper.UserMapperImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
