+++ server/src/main/java/challenge/server/security/filter/JwtVerificationFilter.java
@@ -4,8 +4,8 @@ import challenge.server.exception.BusinessLogicException;
 import challenge.server.exception.ExceptionCode;
 import challenge.server.security.jwt.JwtTokenizer;
 import challenge.server.security.utils.CustomAuthorityUtils;
-import challenge.server.security.user.entity.LogoutList;
-import challenge.server.security.user.service.LogoutListService;
+import challenge.server.user.entity.LogoutList;
+import challenge.server.user.service.LogoutListService;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jws;
