+++ server/src/main/java/challenge/server/security/oauth/service/CustomOAuth2UserService.java
@@ -3,8 +3,8 @@ package challenge.server.security.oauth.service;
 import challenge.server.security.oauth.dto.OAuth2CustomUser;
 import challenge.server.security.oauth.dto.OAuthAttributes;
 import challenge.server.security.utils.CustomAuthorityUtils;
-import challenge.server.security.user.entity.User;
-import challenge.server.security.user.repository.UserRepository;
+import challenge.server.user.entity.User;
+import challenge.server.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.GrantedAuthority;
