+++ server/src/main/java/challenge/server/security/service/SecurityService.java
@@ -1,10 +1,10 @@
 package challenge.server.security.service;
 
 import challenge.server.security.filter.JwtAuthenticationFilter;
-import challenge.server.security.user.dto.UserDto;
-import challenge.server.security.user.mapper.UserMapperImpl;
-import challenge.server.security.user.repository.UserRepository;
-import challenge.server.security.user.service.UserService;
+import challenge.server.user.dto.UserDto;
+import challenge.server.user.mapper.UserMapperImpl;
+import challenge.server.user.repository.UserRepository;
+import challenge.server.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
