+++ server/src/main/java/challenge/server/security/userdetails/CustomUserDetailsService.java
@@ -3,8 +3,8 @@ package challenge.server.security.userdetails;
 import challenge.server.exception.BusinessLogicException;
 import challenge.server.exception.ExceptionCode;
 import challenge.server.security.utils.CustomAuthorityUtils;
-import challenge.server.security.user.entity.User;
-import challenge.server.security.user.repository.UserRepository;
+import challenge.server.user.entity.User;
+import challenge.server.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
