+++ server/src/main/java/challenge/server/user/entity/EmailVerification.java
@@ -1,4 +1,4 @@
-package challenge.server.security.user.entity;
+package challenge.server.user.entity;
 
 import challenge.server.audit.BaseTimeEntity;
 import lombok.*;
