+++ server/src/main/java/challenge/server/user/repository/EmailVerificationRepository.java
@@ -1,6 +1,6 @@
-package challenge.server.security.user.repository;
+package challenge.server.user.repository;
 
-import challenge.server.security.user.entity.EmailVerification;
+import challenge.server.user.entity.EmailVerification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
