+++ server/src/main/java/challenge/server/user/repository/UserCustomRepository.java
@@ -1,6 +1,6 @@
-package challenge.server.security.user.repository;
+package challenge.server.user.repository;
 
-import challenge.server.security.user.dto.UserDto;
+import challenge.server.user.dto.UserDto;
 
 import java.util.List;
