+++ server/src/main/java/challenge/server/user/repository/UserCustomRepositoryImpl.java
