+++ server/src/main/java/challenge/server/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package challenge.server.security.user.repository;
+package challenge.server.user.repository;
 
-import challenge.server.security.user.entity.User;
+import challenge.server.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
